requirements.txt
atomicwrites
attrs
certifi
cffi
charset-normalizer
colorama
coreapi
coreschema
cryptography
defusedxml
django-templated-mail
djoser
python-dotenv==0.19.0
flake8
flake8-quotes2
idna
iniconfig
isort
itypes
Jinja2
MarkupSafe
mccabe
oauthlib
packaging
pluggy
py
pycodestyle
pycparser
pyflakes
pytest
pytest-django
pytest-pythonpath
python-decouple
python3-openid
requests
requests-oauthlib
six
social-auth-app-django
social-auth-core
toml
uritemplate
urllib3
asgiref==3.2.10
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1 